requirements.txt
### CORE
manim==0.19.2
google-generativeai>=0.8.0
python-dotenv>=1.0.0


### UI 
streamlit==1.53.1